+++ packages/react-magma-dom/src/components/Dropdown/DropdownContent.tsx
@@ -67,7 +67,7 @@ export const DropdownContent = React.forwardRef<
   HTMLDivElement,
   DropdownContentProps
 >((props, forwardedRef) => {
-  const { children, testId, ...other } = props;
+  const { children, testId, id, ...other } = props;
   const context = React.useContext(DropdownContext);
   const theme = React.useContext(ThemeContext);
   const ref = useForkedRef(forwardedRef, context.menuRef);
@@ -94,6 +94,7 @@ export const DropdownContent = React.forwardRef<
   return (
     <div
       data-testid={'dropdownContentWrapper'}
+      id={(id ?? context.dropdownButtonId.current) + '_dropdownMenuId'}
       ref={el => {
         context.isOpen && context.setFloating(el);
       }}
@@ -121,8 +122,8 @@ export const DropdownContent = React.forwardRef<
         width={context.width}
       >
         <div
-          aria-labelledby={context.dropdownButtonId.current}
-          role={hasItemChildren ? 'menu' : null}
+          aria-labelledby={id ?? context.dropdownButtonId.current}
+          role={hasItemChildren || hasExpandableItems ? 'menu' : null}
         >
           {children}
         </div>