+++ packages/react-magma-dom/src/components/Select/Select.stories.tsx
@@ -78,6 +78,7 @@ export const Default = {
     isLabelVisuallyHidden: false,
     isMulti: false,
     labelPosition: LabelPosition.top,
+    ariaLabel: 'Another select text',
   },
 };
 
@@ -105,6 +106,7 @@ export const Multi = {
   args: {
     ...Default.args,
     disabled: false,
+    ariaLabel: 'Multi select example',
   },
 };
 
@@ -239,7 +241,7 @@ export const MultiWithDisabledItems = {
       { label: 'Red', value: 'red' },
       { label: 'Green', value: 'green' },
       { label: 'Blue-Disabled', value: 'blue', disabled: true },
-      { label: 'Purple mountain majesty', value: 'purple' },
+      { label: 'Purple mountain majesty', value: 'purple', disabled: false },
       { label: 'Orange', value: 'orange', disabled: false },
       { label: 'Yellow-Disabled', value: 'Yellow-Disabled', disabled: true },
     ],