+++ packages/react-magma-dom/src/components/Tag/Tag.stories.tsx
@@ -110,7 +110,7 @@ const Template: StoryFn<TagProps> = args => {
               console.log('clicked');
             }}
           >
-            Deletetable Tag
+            Deletetable
           </Tag>
         </p>
       </CardBody>
@@ -190,7 +190,7 @@ export const WithDelete = {
           <p>
             {isVisibleDefault && (
               <Tag {...args} onDelete={() => setIsVisibleDefault(false)}>
-                Delete
+                Math
               </Tag>
             )}
           </p>
@@ -201,7 +201,7 @@ export const WithDelete = {
                 onDelete={() => setIsVisibleIcon(false)}
                 icon={<AccountCircleIcon />}
               >
-                Delete Icon
+                Deleteable
               </Tag>
             )}
           </p>
@@ -212,7 +212,7 @@ export const WithDelete = {
                 {...args}
                 onDelete={() => setIsVisibleSmall(false)}
               >
-                Delete Small
+                Deleteable Small
               </Tag>
             )}
           </p>
@@ -224,7 +224,7 @@ export const WithDelete = {
                 onDelete={() => setIsVisibleSmallIcon(false)}
                 icon={<AccountCircleIcon />}
               >
-                Delete Icon Small
+                Icon Small
               </Tag>
             )}
           </p>