+++ packages/react-magma-dom/src/components/Textarea/Textarea.stories.tsx
@@ -72,6 +72,7 @@ export const OnClear = {
       <>
         <Textarea
           {...args}
+          maxCount={4}
           labelText="Textarea"
           value={fieldValue}
           onChange={e => {