+++ website/react-magma-docs/gatsby-config.js
@@ -8,6 +8,8 @@ module.exports = {
   },
   flags: {
     DEV_SSR: false,
+    FAST_DEV: true,
+    PARALLEL_QUERY_RUNNING: true,
   },
   plugins: [
     {