+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Change Log
 
+## 7.0.0-rc.12
+
+### Patch Changes
+
+- e30ddf7: fix(Banner): small fix for docs
+
+## 7.0.0-rc.11
+
+### Patch Changes
+
+- f2b1ce0: chore: Update React-magma-icons version to 3.2.4
+
+## 7.0.0-rc.10
+
+### Patch Changes
+
+- b14bb63: fix(Accordion): Fix bug with not-announced `Expand All/Collapse All` button
+
+## 7.0.0-rc.9
+
+### Patch Changes
+
+- 871d0f2: fix(Nav Tabs): Update documentation to include `aria-labelledby` attribute for multi-instance examples.
+
+## 7.0.0-rc.8
+
+### Patch Changes
+
+- ccb95ae: fix(NavTabs, Tabs, Toast): Fix different issues.
+- d473442: fix(Select): fix issue when scrollbar appears because of padding and negative margin.
+- Updated dependencies [ed37531]
+  - react-magma-dom@5.1.0-rc.13
+
+## 7.0.0-rc.7
+
+### Patch Changes
+
+- 998a3e3: chore: Update React-magma-icons version to 3.2.2
+
 ## 7.0.0-rc.6
 
 ### Patch Changes